Implement CMFD and refactor for Blitz++ use
This one is pretty big, sorry. It adds a lot to make the current workers
correct, and to use those in the CMFD solve, which is now functional and
integrated into the ES. As part of this work, a somewhat pervasive refactor was
done to move a lot of arrays from std::valarray to blitz::Array. Where
appropriate, methods have been changed to take advantage of the Array aliasing
functionality, which may improve performance, but certainly improves readability
(assuming the reader has a decent working knowledge of the Blitz++ library).

+++ src/cmdo/correction_worker.hpp
@@ -45,8 +45,8 @@ namespace mocc {
             {
                 size_t cell_fw = ray.cm_cell_fw();
                 size_t cell_bw = ray.cm_cell_bw();
-                size_t surf_fw = ray.cm_surf_fw();
-                size_t surf_bw = ray.cm_surf_bw();
+                int surf_fw = ray.cm_surf_fw();
+                int surf_bw = ray.cm_surf_bw();
                 size_t iseg_fw = 0;
                 size_t iseg_bw = ray.nseg();
